+++ src/Menu/App/Actions/ProjectModelData.php
@@ -117,16 +117,15 @@ class ProjectModelData
             app()->setLocale($locale); // only way to get localized pagetitle
             $menuItem->setOwnerLabel($resource->getPageTitle($model), $locale);
 
-            if ($model instanceof Visitable && $onlineUrl = $model->url($locale)) {
-
-                $path = Url::fromString($onlineUrl)->getPath();
+            if ($model instanceof Visitable) {
+                $onlineUrl = $model->url($locale);
 
                 /**
                  * For root urls the path is null but in that case we want to have a '/' as path
                  * so we can safely prepend this on each path. However for real missing links
                  * we still set null as url on the menu item.
                  */
-                $menuItem->setUrl($onlineUrl ? '/'.$path : null, $locale);
+                $menuItem->setUrl($onlineUrl ? '/'.Url::fromString($onlineUrl)->getPath() : null, $locale);
             }
         }
 
@@ -144,6 +143,7 @@ class ProjectModelData
             $menuItem->type = MenuLinkType::nolink->value;
             $menuItem->owner_type = null;
             $menuItem->owner_id = null;
+            $menuItem->removeDynamic('url');
             $menuItem->save();
         }
     }